# platform/docs/swagger.yaml
components:
  schemas:
...
    httpx.ErrorResponse:
      properties:
        error:
          type: string
      type: object
...
    httpx.Page-index_SnapshotView:
      properties:
        data:
          items:
            $ref: '#/components/schemas/index.SnapshotView'
          type: array
          uniqueItems: false
        limit:
          type: integer
        page:
          type: integer
        total:
          type: integer
      type: object
...
    index.IndexView:
      properties:
        captured_at:
          example: "2026-07-30T09:15:02Z"
          type: string
        divisor:
          description: |-
            The divisor in force. It is restated whenever an instrument lists, so the
            level does not jump on an event that moved no price.
          example: 1000000000000
          type: number
        kode:
          example: IHSG
          type: string
        market_cap:
          description: |-
            Summed free-float market capitalisation of every priced instrument, in
            rupiah.
          example: 1247830000000000
          format: int64
          type: integer
        members:
          description: How many instruments were priced into this level.
          example: 5
          type: integer
        nama:
          example: Indeks Harga Saham Gabungan
          type: string
        total:
          description: |-
            How many instruments are listed in total. members below this means some
            instrument has no price at all and was excluded rather than counted as zero.
          example: 5
          type: integer
        value:
          description: The index level, rounded to two decimals.
          example: 1247.83
          type: number
      type: object
    index.SnapshotView:
      properties:
        captured_at:
          example: "2026-07-30T09:15:02Z"
          type: string
        divisor:
          example: 1000000000000
          type: number
        market_cap:
          example: 1247830000000000
          format: int64
          type: integer
        members:
          example: 5
          type: integer
        value:
          example: 1247.83
          type: number
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      description: Static administrative key, from configuration. Guards every /api/admin route.
      in: header
      name: X-API-Key
      type: apiKey
    ParticipantKeyAuth:
      description: Per-broker key, issued by an admin and stored hashed. Guards every /api/participant route. Returned once at issue and never retrievable again.
      in: header
      name: X-Participant-Key
      type: apiKey
...
info:
  description: |-
    REST surface for the JAST core matching engine (JATS-style continuous
    matching for Bursa Efek Indonesia). The engine knows only brokers
    (participants), not individual investors, and stops at trade execution.


    There are two authentication tiers, and they never mix:

    - **Participant** (`/api/participant/*`, `/ws/participant/*`) — a broker
    authenticates with its own key, stored hashed in the database and sent as
    `X-Participant-Key`. Keys are issued and revoked per broker at runtime.
    - **Admin** (`/api/admin/*`, `/ws/admin/*`) — the single static key from
    configuration, sent as `X-API-Key`.

    Orders enter only through POST /api/participant/orders. The order book can
    be polled or streamed over WebSocket (documented, but not exercisable from
    Swagger UI, since browsers cannot set headers on a WS handshake).
  title: JAST Core API
  version: 1.0.0
openapi: 3.0.3
paths:
...
  /api/admin/index:
    get:
      description: |-
        Identical to the participant view: the index is one number the whole
        exchange shares, and there is no broker-specific version of it. Offered
        on the admin tier so an operator can read it with the key it already
        holds, rather than having to issue itself a broker key.
      operationId: getAdminIndex
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/index.IndexView'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-API-Key
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Index has not been computed yet
      security:
      - ApiKeyAuth: []
      summary: Read the composite index (admin)
      tags:
      - admin
  /api/admin/index/capture:
    post:
      description: |-
        Writes the current level into the index history immediately, rather than
        waiting for the next automatic capture.

        Useful for marking a specific moment — the end of a session, or the state
        just before an intervention — without depending on where the periodic
        capture happens to fall.

        Returns the level that was recorded. If nothing has been computed yet,
        nothing is written and 503 is returned rather than an empty success.
      operationId: captureIndex
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/index.IndexView'
          description: Created
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-API-Key
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Capture failed
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Index has not been computed yet
      security:
      - ApiKeyAuth: []
      summary: Record a history point now
      tags:
      - admin
  /api/admin/index/history:
    get:
      description: |-
        Identical to the participant view — the same series, offered on the
        admin tier so an operator can read it with the key it already holds.
      operationId: getAdminIndexHistory
      parameters:
      - description: Start of range, RFC 3339
        example: "2026-07-30T00:00:00Z"
        in: query
        name: from
        schema:
          type: string
      - description: End of range, RFC 3339
        example: "2026-07-31T00:00:00Z"
        in: query
        name: to
        schema:
          type: string
      - description: Page number, from 1
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: Rows per page, max 100
        in: query
        name: limit
        schema:
          default: 10
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Page-index_SnapshotView'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: from or to is not RFC 3339
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-API-Key
      security:
      - ApiKeyAuth: []
      summary: Read the composite index history (admin)
      tags:
      - admin
  /api/admin/index/recompute:
    post:
      description: |-
        Revalues the whole market immediately and refreshes the published level,
        instead of waiting for the next trade to trigger it.

        This exists for the case where the level is stale through no fault of the
        market: a price read failed, or an instrument was listed while the index
        could not reach the database. Ordinary operation needs no such call — the
        index recomputes on every execution by itself.

        Admin-only because it is an exchange operation, and because valuing the
        entire market is real work that no broker should be able to schedule at
        will.
      operationId: recomputeIndex
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/index.IndexView'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-API-Key
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Recompute failed
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Nothing could be priced
      security:
      - ApiKeyAuth: []
      summary: Recompute the index now
      tags:
      - admin
...
  /api/participant/index:
    get:
      description: |-
        The exchange-wide price level, computed IHSG-style: the summed
        free-float market capitalisation of every listed instrument, divided by
        a divisor and scaled to a base of 100.

        Each instrument is valued at its last traded price, falling back to its
        IPO price until it has traded — the same reference price the emiten
        detail endpoint reports. An instrument with neither is excluded rather
        than counted as zero, which is why **members** may be lower than
        **total**.

        Weighting is by listed (free-float) shares, not total shares outstanding,
        matching BEI's own methodology since 2021.

        **divisor** is restated whenever an instrument lists, so the level does
        not jump on an event where no price moved. It is returned so a client can
        verify the level rather than trust it.
      operationId: getIndex
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/index.IndexView'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-Participant-Key
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Index has not been computed yet
      security:
      - ParticipantKeyAuth: []
      summary: Read the composite index
      tags:
      - participant
  /api/participant/index/history:
    get:
      description: |-
        Past index levels, newest first. Each point carries the market
        capitalisation and divisor that produced it, so an old level stays
        verifiable even after the divisor has since been restated.

        Optional **from** and **to** bound the range, both RFC 3339 (for example
        `2026-07-30T00:00:00Z`). An unparseable value is rejected with 400 rather
        than silently ignored, since a filter that quietly does nothing returns a
        page the caller will misread as the whole range.
      operationId: getIndexHistory
      parameters:
      - description: Start of range, RFC 3339
        example: "2026-07-30T00:00:00Z"
        in: query
        name: from
        schema:
          type: string
      - description: End of range, RFC 3339
        example: "2026-07-31T00:00:00Z"
        in: query
        name: to
        schema:
          type: string
      - description: Page number, from 1
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: Rows per page, max 100
        in: query
        name: limit
        schema:
          default: 10
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Page-index_SnapshotView'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: from or to is not RFC 3339
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-Participant-Key
      security:
      - ParticipantKeyAuth: []
      summary: Read the composite index history
      tags:
      - participant
...
servers:
- description: Local development
  url: http://localhost:8080